Support for nVidia H200, etc.
If we use `all` or leave it unspecified, the build in `Actions` will crash, so specifying it is necessary. However, in [the previous official build wheels](https://github.com/abetlen/llama-cpp-python/releases/tag/v0.3.4-cu124), it worked even on H200.
This commit seems to resolve the issue for H200 environments for now.

This is just a draft of the fix, and there may be a smarter way to write it.(Such as setting it as an environment variable...) 

It seems that [the installation script for Llama.cpp, which was thought to be one of the causes of the build crash, has been fixed](https://github.com/ggml-org/llama.cpp/pull/14005), but it still doesn't seem to be working properly yet. Specifying it explicitly is still the more reliable option.